beater: move event ACKer from publisher to beater
In the future we will start and stop multiple publishers.
To avoid blocking the publisher.Stop method unnecessarily
during a config reload, move event ACKing to beater.
